+++ js/packages/openinference-trace-exporter/src/vercel/index.ts
@@ -1,4 +1,3 @@
 export * from "./VercelSemanticConventions";
 export * from "./constants";
 export * from "./utils";
-export * from "./types";
